removed browser setup mixin
