+++ lib/browser_factory.js
@@ -10,7 +10,6 @@ const mixins = [
     require("./mixins/browser_info"),
     require("./mixins/browser_navigation"),
     require("./mixins/browser_queries"),
-    require("./mixins/browser_setup"),
     require("./mixins/browser_wait")
 ];
